# docs/surdocs-rest-openapi.yaml
openapi: 3.0.3
info:
  title: SURDOCS External Replacement API
  version: 1.0.0
  description: >
    Proposed REST API contract to replace the legacy SURDOCS_EXTERNAL JSP/JSF
    flows for search, document lookup, document access, and user comments.
servers:
  - url: https://example.dc.gov
...
paths:
...
  /api/v1/search/ssl/options:
    get:
      tags: [Search]
      summary: Get suffix options for a square or parcel-style query
      operationId: getSslOptions
      parameters:
        - in: query
          name: queryType
          required: true
          schema:
            type: string
            enum: [Square, Parcel, Reservation, Appropriation]
        - in: query
          name: square
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Suffix options
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SslOptionsResponse'
  /api/v1/search/ssl/lots:
    get:
      tags: [Search]
      summary: Get lot options for a square/suffix combination
      operationId: getSslLots
      parameters:
        - in: query
          name: sslPrefix
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lot options
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LotOptionsResponse'
...
  /api/v1/search/documents:
    get:
      tags: [Search]
      summary: Search documents by SSL and selected type/subtype
      operationId: searchDocuments
      parameters:
        - in: query
          name: ssl
          required: false
          schema:
            type: string
        - in: query
          name: documentType
          required: false
          schema:
            type: string
        - in: query
          name: subtype
          required: false
          schema:
            type: string
        - in: query
          name: mapNumber
          required: false
          schema:
            type: string
        - in: query
          name: bookName
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Matching documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSearchResponse'
...
components:
  schemas:
...
    SslOptionsResponse:
      type: object
      properties:
        square:
          type: string
        queryType:
          type: string
        suffixes:
          type: array
          items:
            type: string
      required: [square, queryType, suffixes]
    LotOptionsResponse:
      type: object
      properties:
        sslPrefix:
          type: string
        lots:
          type: array
          items:
            type: string
      required: [sslPrefix, lots]
...
    DocumentSearchResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Document'
      required: [items]
    Document:
      type: object
      properties:
        documentId:
          type: string
        guid:
          type: string
        versionSeriesId:
          type: string
        documentType:
          type: string
        subtype:
          type: string
        title:
          type: string
        pageNumber:
          type: string
        mimeType:
          type: string
        category:
          type: string
        sslValues:
          type: array
          items:
            type: string
        mapNumber:
          type: string
        bookName:
          type: string
        contentSize:
          type: integer
          format: int64
      required: [documentId, documentType]
...